Move authorization helpers and error conversions to their own module
This PR only moves code around:
* authorization helpers to `authz_utils`
  * in a subsequent PR, this module will be refactored to no longer
  confuse authentication with authorization
* error mapper to `conversions::api_errors`
