+++ src/internet_identity/src/conversions.rs
@@ -1,0 +1,4 @@
+//! A module to provide type conversions between internal and external types.
+
+/// Type conversion for internal error types to exposed API errors.
+mod api_errors;
